etl-movilidad-local/requirements.txt
# Core dependencies - Google ADK
google-adk>=0.2.0
google-cloud-aiplatform>=1.38.0
pydantic>=2.0.0

# HTTP and feed parsing
requests>=2.31.0
feedparser>=6.0.10
beautifulsoup4>=4.12.0
lxml>=4.9.3

# Web scraping with Apify
apify-client>=1.7.0

# Database - Supabase
supabase>=2.0.0

# Scheduling
schedule>=1.2.0

# Utilities
python-dotenv>=1.0.0
python-dateutil>=2.8.2

# Optional: Email alerts
# smtplib is built-in
